requirements.txt
# ── Core deep learning ────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── Data & numerics ───────────────────────────────────────────────────────────
numpy>=1.24
scipy>=1.11
pandas>=2.0
scikit-learn>=1.3

# ── Financial data ────────────────────────────────────────────────────────────
yfinance>=0.2.36

# ── Image processing ──────────────────────────────────────────────────────────
Pillow>=10.0
matplotlib>=3.7

# ── Progress bars ─────────────────────────────────────────────────────────────
tqdm>=4.65

# ── Notebooks ─────────────────────────────────────────────────────────────────
ipython
ipykernel
jupyter
